requirements.txt
fastapi
uvicorn
motor
pydantic==1.10.9
python-dotenv
redis
passlib[bcrypt]
python-jose